libbpf-cargo: Factor out cargo metadata functions to new module
Will need to reuse this code to implement skeleton generation.
